sensor-board-GUI/requirements.txt
PyQt6
pyserial
argparse